Suppress sync warnings in Django restart
This is a stop-gap measure to allow automatically restarting the server
in development. In production, we don't have this issue, so that's good.
The error appeared to be arising not directly from any code we had
written, and I had no luck isolating it to one component (Sentry, New
Relic, the websocket consumer, and possibly other components were all
candidates), so I opted to suppress the warning and allow restart in
dev, instead.