Add originating user ID to all websocket events
This adds more metadata to every websocket event and will require
matching frontend changes.

Every websocket is now shaped like the error ones have been, with the
model itself in a "model" key, and metadata alongside that key. In this
case, there's now an "originating_user_id" key which is an
Option[String]; if it is None, we are unable to meaningfully attach a
user, and the frontend should probably consider the event relevant to
any user (this applies to creations, deletions, and GitHub webhook
events in particular).

This merits some hand-testing when the frontend changes exist, to
confirm that there aren't possible regressions that the tests missed.