First attempt at solving a11y bug for a long loading message.  (#2072)
I added 'role="status"' to two divs that included the message. There is a
third block that did not wrap the message in a div so I did that. I chose
to use the same pattern as the other blocks using a `React.Fragment` (`<>`)

I created a function that wraps a return result in a div with role='status'

 fixed lint errors; renamed `as_status` to `asStatus`

---------
Co-authored-by: David Ray <david@raycreative.com>